+++ dimos/robot/unitree/go2/dds/msgs/LowState.py
@@ -20,10 +20,10 @@ from dataclasses import dataclass
 
 import numpy as np
 
-from dimos.robot.unitree.go2dds.msgs.base import PrettyMsg
-from dimos.robot.unitree.go2dds.msgs.BmsState import BmsState
-from dimos.robot.unitree.go2dds.msgs.IMUState import IMUState
-from dimos.robot.unitree.go2dds.msgs.MotorState import MotorState
+from dimos.robot.unitree.go2.dds.msgs.base import PrettyMsg
+from dimos.robot.unitree.go2.dds.msgs.BmsState import BmsState
+from dimos.robot.unitree.go2.dds.msgs.IMUState import IMUState
+from dimos.robot.unitree.go2.dds.msgs.MotorState import MotorState
 
 
 @dataclass(repr=False)
