+++ pyproject.toml
@@ -239,7 +239,7 @@ unitree-dds = [
     "dimos[unitree]",
     "unitree-sdk2py-dimos>=1.0.2",
     "cyclonedds>=0.10.5",
-    "mcap>=1.2.0",  # decode Go2 DDS mcap recordings (go2dds store)
+    "mcap>=1.2.0",  # decode Go2 DDS mcap recordings (go2 dds store)
 ]
 
 manipulation = [
@@ -425,7 +425,7 @@ lint = [
 tests-self-hosted = [
     {include-group = "tests"},
     "dimos[agents,perception,manipulation,sim,unitree,misc]",
-    # go2dds store tests decode mcap (pure-Python). The full `unitree-dds` extra
+    # go2 dds store tests decode mcap (pure-Python). The full `unitree-dds` extra
     # pulls `cyclonedds`, whose wheel needs a CycloneDDS C lib the ros-dev image
     # doesn't expose to the build — and these tests never open a live DDS link.
     "mcap>=1.2.0",
